Types & Methods

RootIO.TTree — Type

struct TTree

A struct representing a ROOT TTree with its associated branches and file.

RootIO.TTree — Method

TTree(file::CxxWrap.CxxWrapCore.CxxPtr{ROOT.TFile}, name::String, title::String, data)

Creates a new ROOT TTree with branches of given type or branches having types infered from the given data (data is not written to the tree).

Arguments

file: A pointer to a ROOT file where the TTree will be stored.

name: The name of the TTree.

title: The title of the TTree.

data: The data used to define and optionally fill the branches of the TTree. This can be a DataType or an instance of a type with fields.

Example

mutable struct Event
    x::Float32
    y::Float32
    z::Float32
    v::StdVector{Float32}
end
f = ROOT.TFile!Open("data.root", "RECREATE")
Event()  = Event(0., 0., 0., StdVector{Float32}())
tree = RootIO.TTree(f, "mytree", "mytreetitle", Event)
e = Event()
for i in 1:10
    e.x, e.y, e.z = rand(3)
    resize!.([e.v], 5)
    e.v .= rand(Float32, 5)
    RootIO.Fill(tree, e)
end

RootIO.TTree — Method

TTree(file:CxxWrap.CxxWrapCore.CxxPtr{ROOT.TFile}, name::String, title::String; kwargs...)

Creates a new ROOT TTree and fills it with the provided data.

Arguments

file: A pointer to a ROOT file where the TTree will be stored.

name: The name of the TTree.

title: The title of the TTree.

kwargs...: Named arguments representing the branches of the TTree. Each named argument is either a data type or an array of data.

Example

file = ROOT.TFile!Open("example.root", "RECREATE")
name = "example_tree"
title = "Example TTree"
data = (col_float=rand(Float64, 3), col_int=rand(Int32, 3))
tree = RootIO.TTree(file, name, title; data...)

RootIO.Fill — Method

Fill(tree::TTree, data)

Fills a ROOT TTree with the provided data.

Arguments

tree: The TTree object to be filled.

data: The data to fill the TTree with. This can be a table or a row.

Example

# Assuming `tree` is an existing TTree and `data` is a table or row
Fill(tree, data)

RootIO.Write — Method

Write(tree::TTree)

Writes a ROOT TTree to the associated ROOT file.

Arguments

tree::TTree: The TTree object to be written to the ROOT file.